Implement a new experimental filter
This filter is based on one key observation:

Assuming stationary receivers and transmitters: Given several RSSI
measurements, e.g. -70, -50 and -90 we can deduce that the measurement
with the smallest error must be -50, i.e. the highest observed measurement.
This is because the measured signal strength can never be above the true
signal strength and can only ever get worse due to varying levels of
attenuation or multipath issues.