Add support for Bluetooth extended advertisements
This allows us to restrict advertisements to a specific channel, i.e. 37
in this case. This channel has the lowest frequency and is therefore
usually less affected by path attenuation.

Also, because nodes were previously advertising on two or more channels
the RSSI/distance filter was seeing a lot more jitter because each
Bluetooth channel has different radio characteristics  and usually somewhat
different received signal strengths.

Unfortunately this doesn't solve this problem for non-ESPresense
devices, but it's a first step.